fix: split argmax program factories into separate files
