+++ ttnn/cpp/ttnn/operations/reduction/CMakeLists.txt
@@ -35,7 +35,8 @@ target_sources(
         FILES ${kernels}
     PRIVATE
         argmax/device/argmax_device_operation.cpp
-        argmax/device/argmax_program_factory.cpp
+        argmax/device/argmax_multi_core_program_factory.cpp
+        argmax/device/argmax_single_core_program_factory.cpp
         argmax/argmax.cpp
         accumulation/accumulation_common.cpp
         accumulation/cumprod/cumprod.cpp
