+++ ttnn/cpp/ttnn/operations/reduction/argmax/device/argmax_device_operation.hpp
@@ -5,7 +5,8 @@
 #pragma once
 
 #include "argmax_device_operation_types.hpp"
-#include "argmax_program_factory.hpp"
+#include "argmax_multi_core_program_factory.hpp"
+#include "argmax_single_core_program_factory.hpp"
 #include "ttnn/decorators.hpp"
 #include "ttnn/tensor/tensor.hpp"
