+++ ttnn/cpp/ttnn/operations/reduction/argmax/device/argmax_multi_core_program_factory.hpp
@@ -1,0 +1,35 @@
+// SPDX-FileCopyrightText: © 2025 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#include "argmax_device_operation_types.hpp"
+#include "ttnn/device_operation.hpp"
+
+namespace ttnn::operations::reduction::program {
+
+struct ArgMaxMultiCoreSharedVariables {
+    tt::tt_metal::KernelHandle reader_kernel_id0;
+    tt::tt_metal::KernelHandle reader_kernel_id1;
+    std::vector<CoreCoord> cores_coords0;
+    std::vector<CoreCoord> cores_coords1;
+};
+
+struct ArgMaxMultiCoreProgramFactory {
+    using shared_variables_t = ArgMaxMultiCoreSharedVariables;
+    using cached_program_t = ttnn::device_operation::CachedProgram<shared_variables_t>;
+
+    static cached_program_t create(
+        const operation_attributes_t& operation_attributes,
+        const tensor_args_t& tensor_args,
+        tensor_return_value_t& tensor_return_value);
+
+    static void override_runtime_arguments(
+        cached_program_t& cached_program,
+        const operation_attributes_t& operation_attributes,
+        const tensor_args_t& tensor_args,
+        tensor_return_value_t& tensor_return_value);
+};
+
+}  // namespace ttnn::operations::reduction::program
