+++ ttnn/cpp/ttnn/operations/reduction/argmax/device/argmax_single_core_program_factory.hpp
@@ -30,27 +30,4 @@ struct ArgMaxSingleCoreProgramFactory {
         tensor_return_value_t& tensor_return_value);
 };
 
-struct ArgMaxMultiCoreSharedVariables {
-    tt::tt_metal::KernelHandle reader_kernel_id0;
-    tt::tt_metal::KernelHandle reader_kernel_id1;
-    std::vector<CoreCoord> cores_coords0;
-    std::vector<CoreCoord> cores_coords1;
-};
-
-struct ArgMaxMultiCoreProgramFactory {
-    using shared_variables_t = ArgMaxMultiCoreSharedVariables;
-    using cached_program_t = ttnn::device_operation::CachedProgram<shared_variables_t>;
-
-    static cached_program_t create(
-        const operation_attributes_t& operation_attributes,
-        const tensor_args_t& tensor_args,
-        tensor_return_value_t& tensor_return_value);
-
-    static void override_runtime_arguments(
-        cached_program_t& cached_program,
-        const operation_attributes_t& operation_attributes,
-        const tensor_args_t& tensor_args,
-        tensor_return_value_t& tensor_return_value);
-};
-
 }  // namespace ttnn::operations::reduction::program
